# artifacts/live_staging_test_20250920T160829Z/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "CricAlgo API",
                 "description":  "Cricket Algorithm Trading Bot API",
                 "version":  "0.1.0"
             },
    "paths":  {
...
                  "/api/v1/register":  {
                                           "post":  {
                                                        "tags":  [
                                                                     "authentication"
                                                                 ],
                                                        "summary":  "Register User",
                                                        "description":  "Register a new user and create their wallet.\n\nCreates a user account with optional Telegram ID and returns JWT tokens.",
                                                        "operationId":  "register_user_api_v1_register_post",
                                                        "requestBody":  {
                                                                            "content":  {
                                                                                            "application/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/UserRegister"
                                                                                                                                }
                                                                                                                 }
                                                                                        },
                                                                            "required":  true
                                                                        },
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "Successful Response",
                                                                                      "content":  {
                                                                                                      "application/json":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/TokenResponse"
                                                                                                                                          }
                                                                                                                           }
                                                                                                  }
                                                                                  },
                                                                          "422":  {
                                                                                      "description":  "Validation Error",
                                                                                      "content":  {
                                                                                                      "application/json":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                          }
                                                                                                                           }
                                                                                                  }
                                                                                  }
                                                                      }
                                                    }
                                       },
                  "/api/v1/login":  {
                                        "post":  {
                                                     "tags":  [
                                                                  "authentication"
                                                              ],
                                                     "summary":  "Login User",
                                                     "description":  "Login user with username/password.\n\nFor admin users, TOTP code is required.",
                                                     "operationId":  "login_user_api_v1_login_post",
                                                     "requestBody":  {
                                                                         "content":  {
                                                                                         "application/json":  {
                                                                                                                  "schema":  {
                                                                                                                                 "$ref":  "#/components/schemas/UserLogin"
                                                                                                                             }
                                                                                                              }
                                                                                     },
                                                                         "required":  true
                                                                     },
                                                     "responses":  {
                                                                       "200":  {
                                                                                   "description":  "Successful Response",
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/TokenResponse"
                                                                                                                                       }
                                                                                                                        }
                                                                                               }
                                                                               },
                                                                       "422":  {
                                                                                   "description":  "Validation Error",
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                       }
                                                                                                                        }
                                                                                               }
                                                                               }
                                                                   }
                                                 }
                                    },
                  "/api/v1/refresh":  {
                                          "post":  {
                                                       "tags":  [
                                                                    "authentication"
                                                                ],
                                                       "summary":  "Refresh Token",
                                                       "description":  "Refresh access token using refresh token.",
                                                       "operationId":  "refresh_token_api_v1_refresh_post",
                                                       "requestBody":  {
                                                                           "content":  {
                                                                                           "application/json":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/RefreshTokenRequest"
                                                                                                                               }
                                                                                                                }
                                                                                       },
                                                                           "required":  true
                                                                       },
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "Successful Response",
                                                                                     "content":  {
                                                                                                     "application/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/TokenResponse"
                                                                                                                                         }
                                                                                                                          }
                                                                                                 }
                                                                                 },
                                                                         "422":  {
                                                                                     "description":  "Validation Error",
                                                                                     "content":  {
                                                                                                     "application/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                         }
                                                                                                                          }
                                                                                                 }
                                                                                 }
                                                                     }
                                                   }
                                      },
                  "/api/v1/me":  {
                                     "get":  {
                                                 "tags":  [
                                                              "authentication"
                                                          ],
                                                 "summary":  "Get Current User Info",
                                                 "description":  "Get current user information.",
                                                 "operationId":  "get_current_user_info_api_v1_me_get",
                                                 "responses":  {
                                                                   "200":  {
                                                                               "description":  "Successful Response",
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {

                                                                                                                                   }
                                                                                                                    }
                                                                                           }
                                                                           }
                                                               },
                                                 "security":  [
                                                                  {
                                                                      "HTTPBearer":  [

                                                                                     ]
                                                                  }
                                                              ]
                                             }
                                 },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
                                       "RefreshTokenRequest":  {
                                                                   "properties":  {
                                                                                      "refresh_token":  {
                                                                                                            "type":  "string",
                                                                                                            "title":  "Refresh Token"
                                                                                                        }
                                                                                  },
                                                                   "type":  "object",
                                                                   "required":  [
                                                                                    "refresh_token"
                                                                                ],
                                                                   "title":  "RefreshTokenRequest",
                                                                   "description":  "Refresh token request model"
                                                               },
...
                                       "TokenResponse":  {
                                                             "properties":  {
                                                                                "access_token":  {
                                                                                                     "type":  "string",
                                                                                                     "title":  "Access Token"
                                                                                                 },
                                                                                "refresh_token":  {
                                                                                                      "type":  "string",
                                                                                                      "title":  "Refresh Token"
                                                                                                  },
                                                                                "token_type":  {
                                                                                                   "type":  "string",
                                                                                                   "title":  "Token Type",
                                                                                                   "default":  "bearer"
                                                                                               },
                                                                                "expires_in":  {
                                                                                                   "type":  "integer",
                                                                                                   "title":  "Expires In"
                                                                                               }
                                                                            },
                                                             "type":  "object",
                                                             "required":  [
                                                                              "access_token",
                                                                              "refresh_token",
                                                                              "expires_in"
                                                                          ],
                                                             "title":  "TokenResponse",
                                                             "description":  "Token response model"
                                                         },
...
                                       "UserLogin":  {
                                                         "properties":  {
                                                                            "username":  {
                                                                                             "type":  "string",
                                                                                             "title":  "Username"
                                                                                         },
                                                                            "password":  {
                                                                                             "type":  "string",
                                                                                             "title":  "Password"
                                                                                         },
                                                                            "totp_code":  {
                                                                                              "anyOf":  [
                                                                                                            {
                                                                                                                "type":  "string"
                                                                                                            },
                                                                                                            {
                                                                                                                "type":  "null"
                                                                                                            }
                                                                                                        ],
                                                                                              "title":  "Totp Code"
                                                                                          }
                                                                        },
                                                         "type":  "object",
                                                         "required":  [
                                                                          "username",
                                                                          "password"
                                                                      ],
                                                         "title":  "UserLogin",
                                                         "description":  "User login request model"
                                                     },
                                       "UserRegister":  {
                                                            "properties":  {
                                                                               "username":  {
                                                                                                "type":  "string",
                                                                                                "maxLength":  48,
                                                                                                "minLength":  3,
                                                                                                "title":  "Username"
                                                                                            },
                                                                               "telegram_id":  {
                                                                                                   "anyOf":  [
                                                                                                                 {
                                                                                                                     "type":  "integer"
                                                                                                                 },
                                                                                                                 {
                                                                                                                     "type":  "null"
                                                                                                                 }
                                                                                                             ],
                                                                                                   "title":  "Telegram Id"
                                                                                               },
                                                                               "password":  {
                                                                                                "anyOf":  [
                                                                                                              {
                                                                                                                  "type":  "string",
                                                                                                                  "minLength":  8
                                                                                                              },
                                                                                                              {
                                                                                                                  "type":  "null"
                                                                                                              }
                                                                                                          ],
                                                                                                "title":  "Password"
                                                                                            }
                                                                           },
                                                            "type":  "object",
                                                            "required":  [
                                                                             "username"
                                                                         ],
                                                            "title":  "UserRegister",
                                                            "description":  "User registration request model"
                                                        },
...
                                   },
                       "securitySchemes":  {
                                               "HTTPBearer":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer"
                                                              }
                                           }
                   }
}